ecohotel_board/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
bitarray==2.7.4
cached-property==1.5.2
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cytoolz==0.12.1
Django==3.2.19
django-crispy-forms==2.0
django-redis==5.2.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
hiredis==2.2.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
parsimonious==0.9.0
pkgutil-resolve-name==1.3.10
protobuf==4.22.3
pycryptodome==3.17
pyrsistent==0.19.3
pytz==2023.3
pywin32==306
PyYAML==6.0
redis==4.5.4
regex==2023.5.5
requests==2.29.0
rlp==3.0.0
six==1.16.0
sqlparse==0.4.4
toolz==0.12.0
typing-extensions==4.5.0
urllib3==1.26.15
web3==6.3.0
websockets==11.0.2
yarl==1.9.2
zipp==3.15.0
